+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - restore annotation `@DtoUpdateHidden`
 - output NestJs pluralized model folders.
 - add `@ApiProperty`.type for `@DtoCastType'
+- support nested parentheses for validators, e.g.: @Matches(/^[a-z\d](?:[a-z\d]|-(?=[a-z\d])){0,38}$/i)
 
 ## [1.17.4] - 2023-03-04